tool/pyproject.toml
[project]
name = "openai-tools"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Hamza", email = "hamzaamir7651@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "chainlit>=2.5.5",
    "openai-agents>=0.0.15",
]

[project.scripts]
openai-tools = "openai_tools:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
